requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
Jinja2==3.1.2
pytest==7.4.3
flake8==6.1.0
black==23.10.1
Flask-Dance==7.0.0
blinker==1.6.2
requests==2.31.0
oauthlib==3.2.2
